backend/requirements.txt
fastapi
uvicorn
sqlalchemy
sqlmodel
bcrypt
python-jose[cryptography]
python-multipart
twilio
python-dotenv
pandas
plotly
